Various cleaning, improvement and reorganisation (WIP)
