+++ til_base_model/_Demo/_migration/migration.yml
@@ -13,12 +13,12 @@ entities:
                 - select_men: logit_regr(
                     0.0,
                     filter= ISMALE,
-                    align = 'param/demo/hyp_soldemigH_custom.csv'
+                    align = 'parameters/population/hyp_soldemigH_custom.csv'
                     )
                 - select_women: logit_regr(
                     0.0,
                     filter= ISFEMALE,
-                    align = 'param/demo/hyp_soldemigF_custom.csv'
+                    align = 'parameters/population/hyp_soldemigF_custom.csv'
                     )
                 - select: select_men or select_women
                 - new('individus', filter = select,
@@ -34,9 +34,9 @@ entities:
                     choi = 0,
                     migrant = True,
                     quimen = 0,
-                    quifoy = 0
+                    quifoy = 0,
+                    dur_in_couple = 0
                     )
-                    ,
                     # select = 2
                     # naiss = naiss
                 - csv(
